[Feat] adding BadRequestException , improving inventory service.

+++ InventoryManagementSystem.Core/Entities/Product/Product.cs
@@ -8,7 +8,7 @@ namespace InventoryManagementSystem.Core.Entities.Product
 {
     public class Product
     {
-        public decimal ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public decimal ProductPrice { get; set; }
