+++ InventoryManagementSystem.Core/Interfaces/Repositories/AllProductIRepository/IProductRepository.cs
@@ -11,7 +11,7 @@ namespace InventoryManagementSystem.Core.Interfaces.Repositories.AllProductIRepo
     {
         Task<Product> GetProduct(int ProductId);
         Task<List<Product>> GetProducts();
-        Task<decimal> InsertProduct(Product Product);
+        Task<int> InsertProduct(Product Product);
         Task DeleteProduct(int ProductId);
     }
 }
